Fixed serializing filter value

+++ src/main/kotlin/co/llective/hyena/api/MessageBuilder.kt
@@ -213,13 +213,13 @@ object MessageBuilder {
     private fun writeFilterValue(dos: DataOutput, filter: ScanFilter) {
         dos.writeInt(filter.type.ordinal)
         when (filter.type) {
-            FilterType.I8 -> dos.writeByte(filter.value as Int)
-            FilterType.I16 -> dos.writeShort(filter.value as Int)
-            FilterType.I32 -> dos.writeInt(filter.value as Int)
+            FilterType.I8 -> dos.writeByte((filter.value as Long).toInt())
+            FilterType.I16 -> dos.writeShort((filter.value as Long).toInt())
+            FilterType.I32 -> dos.writeInt((filter.value as Long).toInt())
             FilterType.I64 -> dos.writeLong(filter.value as Long)
 
-            FilterType.U8 -> dos.writeShort(filter.value as Int)
-            FilterType.U16 -> dos.writeInt(filter.value as Int)
+            FilterType.U8 -> dos.writeShort((filter.value as Long).toInt())
+            FilterType.U16 -> dos.writeInt((filter.value as Long).toInt())
             FilterType.U32 -> dos.writeLong(filter.value as Long)
             FilterType.U64 -> writeU64(dos, BigInteger.valueOf(filter.value as Long))
